discord: Util::Logging を IRC アダプター専用に戻した
