+++ lib/rgrb/plugin/util/logging.rb
@@ -4,28 +4,40 @@ module RGRB
   module Plugin
     module Util
       # プラグインのログ記録関連メソッドを集めたモジュール。
+      # IRC アダプター専用
       module Logging
         private
 
-        def self.included(by)
-          name = by.to_s.split('::').last
-          require "rgrb/plugin/util/logging-#{name.to_snake}"
+        # 入ってきたメッセージをログに残す
+        # @param [Cinch::Message] m メッセージ
+        # @return [void]
+        def log_incoming(m)
+          log(m.raw, :incoming, :info)
+        end
+
+        # NOTICE をログに残す
+        # @param [Cinch::Target] target NOTICE の対象
+        # @param [String] message メッセージ
+        # @return [void]
+        def log_notice(target, message)
+          log("<NOTICE to #{target.name}> #{message.inspect}", :outgoing, :info)
+        end
+
+        # JOIN をログに残す
+        # @param [Cinch::Channel] channel
+        # @return [void]
+        def log_join(channel)
+          log("<JOIN on #{channel}>", :outgoing, :info)
+        end
+
+        # PART をログに残す
+        # @param [Cinch::Channel] channel
+        # @param [String] message 退出メッセージ
+        # @return [void]
+        def log_part(channel, message)
+          log("<PART from #{channel}> #{message.inspect}", :outgoing, :info)
         end
       end
     end
-  end
-end
-
-class String
-  def to_snake
-    self
-      .gsub(/([A-Z]+)([A-Z][a-z])/, '\1_\2')
-      .gsub(/([a-z\d])([A-Z])/, '\1_\2')
-      .tr('-', '_')
-      .downcase
-  end
-
-  def to_camel
-    self.split('_').map {|w| w[0] = w[0].upcase; w }.join
   end
 end
